backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend with Supabase"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "supabase>=2.3.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic[email]>=2.5.3",
    "pydantic-settings>=2.1.0",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "asyncpg>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "httpx>=0.26.0",
    "ruff>=0.1.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
